add optional onLogicInit function called after async load
In case you need to do any processing immediately after loading logic, if you export a onLogicInit function it will be called with the store.

You can use this to dispatch actions immediately after load. It will also be called each time you navigate back to this container but by default the code checks to see if it is the same store and only runs once (while still allowing mulitple runs for tests).